+++ tests/testthat/test-bias_correction.R
@@ -32,7 +32,12 @@ test_that("ggpredict, bias_correction, mixed", {
     data = dat,
     family = binomial(link = "logit")
   )
-  out1 <- predict_response(m1, "var_binom")
+  expect_message(
+    {
+      out1 <- predict_response(m1, "var_binom")
+    },
+    regex = "You are calculating"
+  )
   out2 <- predict_response(m1, "var_binom", bias_correction = TRUE)
   out3 <- as.data.frame(emmeans::emmeans(m1, "var_binom", type = "response"))
   out4 <- as.data.frame(emmeans::emmeans(
@@ -45,6 +50,9 @@ test_that("ggpredict, bias_correction, mixed", {
 
   expect_equal(out1$predicted, out3$prob, tolerance = 1e-3)
   expect_equal(out2$predicted, out4$prob, tolerance = 1e-3)
+
+  # no message when type = "random"
+  expect_silent(predict_response(m1, "var_binom", type = "random"))
 })
 
 skip_if_not_installed("glmmTMB")
@@ -66,7 +74,12 @@ test_that("ggpredict, bias_correction, glmmTMB", {
     bias.adjust = TRUE,
     sigma = sqrt(insight::get_variance_residual(m3))
   ))
-  out3 <- predict_response(m3, "mined", margin = "marginalmeans")
+  expect_message(
+    {
+      out3 <- predict_response(m3, "mined", margin = "marginalmeans")
+    },
+    regex = "You are calculating"
+  )
   out4 <- predict_response(m3, "mined", margin = "marginalmeans", bias_correction = TRUE)
   expect_equal(out1$rate, out3$predicted, tolerance = 1e-3)
   expect_equal(out2$rate, out4$predicted, tolerance = 1e-3)
@@ -77,7 +90,7 @@ test_that("ggpredict, bias_correction, glmmTMB", {
   expect_equal(result1$Contrast, -1.84456, tolerance = 1e-3)
   expect_equal(result2$Contrast, -2.11459, tolerance = 1e-3)
 
-  out5 <- predict_response(m3, "mined")
+  out5 <- predict_response(m3, "mined", verbose = FALSE)
   expect_equal(out5$predicted, c(0.93517, 2.57911), tolerance = 1e-3)
   out6 <- predict_response(m3, "mined", bias_correction = TRUE, sigma = sqrt(insight::get_variance_residual(m3)))
   expect_equal(out6$predicted, c(1.07207, 2.95666), tolerance = 1e-3)