+++ NEWS.md
@@ -1,4 +1,4 @@
-# ggeffects (development)
+# ggeffects 1.8.0
 
 ## Breaking changes
 
@@ -11,6 +11,11 @@
   `ggemmeans()` have been removed.
 
 * The deprecated arguments for the `plot()` method have been removed.
+
+## Changes
+
+* `test_predictions` gains a `test_args` argument, to optionally pass further
+  options to `test` for *emmeans* options.
 
 # ggeffects 1.7.2